+++ packages/headless/doc-parser/doc-parser.ts
@@ -11,6 +11,17 @@ const entryPoint = apiPackage.entryPoints[0];
 
 const controllers: ControllerConfiguration[] = [
   {
+    initializer: 'buildFacet',
+    samplePaths: {
+      react_class: [
+        'packages/samples/headless-react/src/components/facet/facet.class.tsx',
+      ],
+      react_fn: [
+        'packages/samples/headless-react/src/components/facet/facet.fn.tsx',
+      ],
+    },
+  },
+  {
     initializer: 'buildPager',
     samplePaths: {
       react_class: [
@@ -28,6 +39,10 @@ const controllers: ControllerConfiguration[] = [
         'packages/samples/headless-react/src/components/context/context.ts',
       ],
     },
+  },
+  {
+    initializer: 'buildCategoryFacet',
+    samplePaths: {},
   },
   {
     initializer: 'buildDidYouMean',
@@ -81,6 +96,37 @@ const controllers: ControllerConfiguration[] = [
       ],
       react_fn: [
         'packages/samples/headless-react/src/components/query-summary/query-summary.fn.tsx',
+      ],
+    },
+  },
+  {
+    initializer: 'buildSearchStatus',
+    samplePaths: {
+      react_class: [
+        'packages/samples/headless-react/src/components/search-status/search-status.class.tsx',
+      ],
+      react_fn: [
+        'packages/samples/headless-react/src/components/search-status/search-status.fn.tsx',
+      ],
+    },
+  },
+  {
+    initializer: 'buildInteractiveResult',
+    samplePaths: {
+      react_class: [],
+      react_fn: [
+        'packages/samples/headless-react/src/components/result-list/result-link.tsx',
+      ],
+    },
+  },
+  {
+    initializer: 'buildResultsPerPage',
+    samplePaths: {
+      react_class: [
+        'packages/samples/headless-react/src/components/results-per-page/results-per-page.class.tsx',
+      ],
+      react_fn: [
+        'packages/samples/headless-react/src/components/results-per-page/results-per-page.fn.tsx',
       ],
     },
   },