+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,37 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.7.0-alpha.25](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.24...v0.7.0-alpha.25) (2021-03-01)
+
+
+### Features
+
+* **atomic:** make default CSS variables available at the root ([#551](https://github.com/coveo/ui-kit/issues/551)) ([950c1fc](https://github.com/coveo/ui-kit/commit/950c1fc58cbbcebc06265fc66d6c79c39158271e))
+* **atomic:** remove atomic-result-excerpt component from DOM when empty ([#550](https://github.com/coveo/ui-kit/issues/550)) ([28baaae](https://github.com/coveo/ui-kit/commit/28baaae439c9dbf5d7df139013615b57c2d99733))
+
+
+
+
+
+# [0.7.0-alpha.24](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.23...v0.7.0-alpha.24) (2021-02-26)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [0.7.0-alpha.23](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.22...v0.7.0-alpha.23) (2021-02-26)
+
+
+### Features
+
+* **atomic:** facet Sort Criterion and Enable Facet Search Prop ([#543](https://github.com/coveo/ui-kit/issues/543)) ([a2c50c3](https://github.com/coveo/ui-kit/commit/a2c50c357144a024a7dca551ada59f4e72c6f9ff))
+
+
+
+
+
 # [0.7.0-alpha.22](https://github.com/coveo/ui-kit/compare/v0.7.0-alpha.21...v0.7.0-alpha.22) (2021-02-26)